+++ sam/data_sources/__init__.py
@@ -1,12 +1,13 @@
 from .synthetic_data import synthetic_timeseries, synthetic_date_range
-from .knmy_stations import knmy_stations
-from .weather import read_knmi, read_openweathermap, read_regenradar
 from .mongo_wrapper import MongoWrapper
 from .deprecated import create_synthetic_timeseries
+from .weather.knmi import read_knmi, read_knmi_station_data, read_knmi_stations
+from .weather.openweathermap import read_openweathermap
+from .weather.regenradar import read_regenradar
 
-from . import mongo_wrapper
-from . import synthetic_data
-from . import weather
 
-__all__ = ["synthetic_timeseries", "synthetic_date_range", "read_knmi",
-           "read_openweathermap", "read_regenradar", "MongoWrapper"]
+__all__ = ["synthetic_timeseries", "synthetic_date_range", "create_synthetic_timeseries",
+           "read_knmi", "read_knmi_station_data", "read_knmi_stations",
+           "read_openweathermap",
+           "read_regenradar",
+           "MongoWrapper"]